+++ src/views/text-numeric.js
@@ -1,13 +1,9 @@
+import { numDelim } from '../core/utils/html.js';
 import usage from './text-numeric.usage.js';
 
 export default function(host) {
     host.view.define('text-numeric', function (el, config, data) {
-        const value = String(data);
-
-        el.innerHTML = value.replace(
-            /\.\d+(eE[-+]?\d+)?|\B(?=(\d{3})+(\D|$))/g,
-            m => m || '<span class="num-delim"></span>'
-        );
+        el.innerHTML = numDelim(data);
     }, {
         tag: 'span',
         usage